requirements.txt
python-dotenv
openai
diskcache
pyyaml
ipykernel
google-generativeai
